+++ common/src/main/java/com/faboslav/structurify/world/level/structure/StructureSectionClaim.java
@@ -1,6 +1,3 @@
 package com.faboslav.structurify.world.level.structure;
 
-// Maybe just keep the long, and create separate debug data
-public record StructureSectionClaim(long token, String structureId, String pieceId)
-{
-}
+public record StructureSectionClaim(long token, String structureId, long structureCenter) {}